joyeuse: Remove AOSP prefix
This is how Google builds the Pixel stock ROM, and it makes more sense
in my opinion: this is generally an Android device tree. Instead of
branding the device differently in every custom ROM despite there being
few actual differences, remove.

Note that this has been added as a new product that overlays the AOSP
product to avoid merge conflicts in AOSP updates from Google and ease
commit exchanging with other ROMs based on AOSP trees.

Change-Id: Ie3c401fd6e32dc7758b5590f9968ff3787687ebe
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>