joyeuse: overlay: only use proximity sensor if supported
On msm-4.14 devices, when the proximity sensor is in use,
the smp2p-sleepstate IRQ is fired multiple times a second,
with each one holding a 200ms wakelock.
This is probably a bug in the DSP firmware.
To fix this, avoid using the proximity sensor in doze mode,
because sleep is preferred to turning off the screen.

Change-Id: I57750afd77267abdc49780f70636626d20e666ad
Signed-off-by: saikiran2001 <bjsaikiran@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: ihsan <ihsanulrahman@gmail.com>